Set up connection between microservice order and inventory. Created new Interface to update Quantity in the inventory services, whenever new order is created.

+++ clients/src/main/java/com/bartoszmarkiewicz/clients/inventory/ProductResponse.java
@@ -2,6 +2,6 @@ package com.bartoszmarkiewicz.clients.inventory;
 
 public record ProductResponse(
         Integer productId,
-        Float productQuantity
+        Integer productQuantity
 ) {
 }
